Implementation java and jar files - sinh Problem 4
These files are being created during the implementation phase. The SinhLogic .java is the class responsible for the business logic and the SinhUI.java is responsible for the User Interface, the jar file is the application file that covers both java files. These all files together compose the solution implemented for the resolution of the function sinh Problem 4.